refactor(bench): Consolidate with examples
