+++ examples/ini/parser_str.rs
@@ -9,7 +9,9 @@ use winnow::{
   sequence::{delimited, terminated},
 };
 
-pub fn categories(input: &str) -> IResult<&str, HashMap<&str, HashMap<&str, &str>>> {
+pub type Input<'i> = &'i str;
+
+pub fn categories(input: Input<'_>) -> IResult<Input<'_>, HashMap<&str, HashMap<&str, &str>>> {
   match categories_aggregator(input) {
     Ok((i, tuple_vec)) => Ok((i, tuple_vec.into_iter().collect())),
     Err(e) => Err(e),
@@ -17,33 +19,33 @@ pub fn categories(input: &str) -> IResult<&str, HashMap<&str, HashMap<&str, &str
 }
 
 #[allow(clippy::type_complexity)]
-fn categories_aggregator(i: &str) -> IResult<&str, Vec<(&str, HashMap<&str, &str>)>> {
+fn categories_aggregator(i: Input<'_>) -> IResult<Input<'_>, Vec<(&str, HashMap<&str, &str>)>> {
   many0(category_and_keys)(i)
 }
 
-fn category_and_keys(i: &str) -> IResult<&str, (&str, HashMap<&str, &str>)> {
+fn category_and_keys(i: Input<'_>) -> IResult<Input<'_>, (&str, HashMap<&str, &str>)> {
   (category, keys_and_values).parse_next(i)
 }
 
-fn category(i: &str) -> IResult<&str, &str> {
+fn category(i: Input<'_>) -> IResult<Input<'_>, &str> {
   terminated(
     delimited('[', take_while(|c| c != ']'), ']'),
     opt(take_while1(" \r\n")),
   )(i)
 }
 
-fn keys_and_values(input: &str) -> IResult<&str, HashMap<&str, &str>> {
+fn keys_and_values(input: Input<'_>) -> IResult<Input<'_>, HashMap<&str, &str>> {
   match keys_and_values_aggregator(input) {
     Ok((i, tuple_vec)) => Ok((i, tuple_vec.into_iter().collect())),
     Err(e) => Err(e),
   }
 }
 
-fn keys_and_values_aggregator(i: &str) -> IResult<&str, Vec<(&str, &str)>> {
+fn keys_and_values_aggregator(i: Input<'_>) -> IResult<Input<'_>, Vec<(&str, &str)>> {
   many0(key_value)(i)
 }
 
-fn key_value(i: &str) -> IResult<&str, (&str, &str)> {
+fn key_value(i: Input<'_>) -> IResult<Input<'_>, (&str, &str)> {
   let (i, key) = alphanumeric(i)?;
   let (i, _) = (opt(space), "=", opt(space)).parse_next(i)?;
   let (i, val) = take_till(is_line_ending_or_comment)(i)?;
@@ -58,11 +60,11 @@ fn is_line_ending_or_comment(chr: char) -> bool {
   chr == ';' || chr == '\n'
 }
 
-fn not_line_ending(i: &str) -> IResult<&str, &str> {
+fn not_line_ending(i: Input<'_>) -> IResult<Input<'_>, &str> {
   take_while(|c| c != '\r' && c != '\n')(i)
 }
 
-fn space_or_line_ending(i: &str) -> IResult<&str, &str> {
+fn space_or_line_ending(i: Input<'_>) -> IResult<Input<'_>, &str> {
   take_while1(" \r\n")(i)
 }
