+++ examples/json/parser.rs
@@ -209,8 +209,10 @@ fn ws<'i, E: ParseError<&'i str>>(input: &'i str) -> IResult<&'i str, &'i str, E
 
 #[cfg(test)]
 mod test {
+  #[allow(dead_code)] // its dead for benches
   use super::*;
 
+  #[allow(dead_code)] // its dead for benches
   type Error<'i> = winnow::error::Error<&'i str>;
 
   #[test]
